requirements.txt
autopep8==1.5.6
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.7
future==0.18.2
netmiko==3.3.3
ntc-templates==2.0.0
paramiko==2.7.2
pycodestyle==2.7.0
pycparser==2.20
PyNaCl==1.4.0
pyserial==3.5
python-dotenv==0.17.0
scp==0.13.3
six==1.15.0
tenacity==7.0.0
textfsm==1.1.0
toml==0.10.2
